Arrays now stored in FLoat32 precision to reduce read/write while computation remains at Float64 precision + various small optimisations

+++ src/Binary/MomentumSpaceFactors.jl
@@ -45,17 +45,17 @@ end
 
 Applies momentum space volume element to the Chunked Gain and Loss matrices such that they have the correct dimensions for use in particle transport, i.e. they have units of length^3/time (normalised by σT*c)
 """
-function MomentumSpaceFactorsBinaryChunk!(ChunkGainMatrix3::Array{Float64,7},ChunkGainMatrix4::Array{Float64,7},u3_bounds::Vector{Float64},h3_bounds::Vector{Float64},u4_bounds::Vector{Float64},h4_bounds::Vector{Float64},Indistinguishable_12::Bool)
+function MomentumSpaceFactorsBinaryChunk!(ChunkGainMatrix3::Array{Float32,7},ChunkGainMatrix4::Array{Float32,7},u3_bounds::Vector{Float64},h3_bounds::Vector{Float64},u4_bounds::Vector{Float64},h4_bounds::Vector{Float64},Indistinguishable_12::Bool)
 
     # Momentum space volume elements
     for h2 in axes(ChunkGainMatrix3,7), u2 in axes(ChunkGainMatrix3,6), h1 in axes(ChunkGainMatrix3,5), u1 in axes(ChunkGainMatrix3,4) # common axes
         for h3 in axes(ChunkGainMatrix3,3), u3 in axes(ChunkGainMatrix3,2), p3 in 1:size(ChunkGainMatrix3,1)
-            ChunkGainMatrix3[p3,u3,h3,u1,h1,u2,h2] *= (u3_bounds[u3+1]-u3_bounds[u3])*(h3_bounds[h3+1]-h3_bounds[h3]) # du3dh3
-            ChunkGainMatrix3[p3,u3,h3,u1,h1,u2,h2] /= (1e0+Float64(Indistinguishable_12))
+            ChunkGainMatrix3[p3,u3,h3,u1,h1,u2,h2] *= Float32((u3_bounds[u3+1]-u3_bounds[u3])*(h3_bounds[h3+1]-h3_bounds[h3])) # du3dh3
+            ChunkGainMatrix3[p3,u3,h3,u1,h1,u2,h2] /= Float32(1+Indistinguishable_12)
         end
         for h4 in axes(ChunkGainMatrix4,3), u4 in axes(ChunkGainMatrix4,2), p4 in 1:size(ChunkGainMatrix4,1)
-            ChunkGainMatrix4[p4,u4,h4,u1,h1,u2,h2] *= (u4_bounds[u4+1]-u4_bounds[u4])*(h4_bounds[h4+1]-h4_bounds[h4]) # du4dh4
-            ChunkGainMatrix4[p4,u4,h4,u1,h1,u2,h2] /= (1e0+Float64(Indistinguishable_12))
+            ChunkGainMatrix4[p4,u4,h4,u1,h1,u2,h2] *= Float32((u4_bounds[u4+1]-u4_bounds[u4])*(h4_bounds[h4+1]-h4_bounds[h4])) # du4dh4
+            ChunkGainMatrix4[p4,u4,h4,u1,h1,u2,h2] /= Float32(1+Indistinguishable_12)
         end
     end
 
@@ -63,13 +63,13 @@ function MomentumSpaceFactorsBinaryChunk!(ChunkGainMatrix3::Array{Float64,7},Chu
 
 end
 
-function MomentumSpaceFactorsBinaryChunk!(ChunkGainMatrix3::Array{Float64,7},u3_bounds::Vector{Float64},h3_bounds::Vector{Float64},Indistinguishable_12::Bool)
+function MomentumSpaceFactorsBinaryChunk!(ChunkGainMatrix3::Array{Float32,7},u3_bounds::Vector{Float64},h3_bounds::Vector{Float64},Indistinguishable_12::Bool)
 
     # Momentum space volume elements
     for h2 in axes(ChunkGainMatrix3,7), u2 in axes(ChunkGainMatrix3,6), h1 in axes(ChunkGainMatrix3,5), u1 in axes(ChunkGainMatrix3,4) # common axes
         for h3 in axes(ChunkGainMatrix3,3), u3 in axes(ChunkGainMatrix3,2), p3 in 1:size(ChunkGainMatrix3,1)
-            ChunkGainMatrix3[p3,u3,h3,u1,h1,u2,h2] *= (u3_bounds[u3+1]-u3_bounds[u3])*(h3_bounds[h3+1]-h3_bounds[h3]) # du3dh3
-            ChunkGainMatrix3[p3,u3,h3,u1,h1,u2,h2] /= (1e0+Float64(Indistinguishable_12))
+            ChunkGainMatrix3[p3,u3,h3,u1,h1,u2,h2] *= Float32((u3_bounds[u3+1]-u3_bounds[u3])*(h3_bounds[h3+1]-h3_bounds[h3])) # du3dh3
+            ChunkGainMatrix3[p3,u3,h3,u1,h1,u2,h2] /= Float32(1+Indistinguishable_12)
         end
     end
